+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- `CodeSpan` support
+
 ## [0.11.0] - 2019-09-16
 
 ### Added